Added additional tests for basic cases.

+++ NewsFeed/Data/Australia/NSW/BomData.cs
@@ -6,7 +6,7 @@
         public string name { get; set; }
         public string local_date_time { get; set; }
         public long aifstime_utc { get; set; }
-        public float air_temp { get; set; }
+        public double air_temp { get; set; }
         public string wind_dir { get; set; }
         public int wind_spd_kmh { get; set; }
     }
